{EE-789 / EE-790} address some comments

+++ execution-engine/contract-ffi/src/value/account.rs
@@ -213,7 +213,7 @@ impl Debug for PublicKey {
 
 impl CLTyped for PublicKey {
     fn cl_type() -> CLType {
-        CLType::FixedList(Box::new(CLType::U8), KEY_SIZE as u64)
+        CLType::FixedList(Box::new(CLType::U8), KEY_SIZE as u32)
     }
 }
